Version with finished Tic-Tac-Toe
Here is a new version of our project with tic-tac-toe done! i was able to figure out how to make it so the data for tic-tac-toe is stored in a class of it's own rather than the class created by the form. If you guys have any questions feel free to ask!!!